CI/travis: use official Qwt package
For Darwin, we need to build it, but at least we're downloading it from the
official site (on SourceForge) now.

For Ubuntu, we also need to build it for versions prior to 18.04.
That's because we're also using a unofficial Qt5 package (to get version
5.9), so it's a good idea to build.

Ubuntu 18.04 has Qt 5.9.5 packaged, so we can use that directly.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>